requirements.txt
absl-py==0.13.0
addict==2.4.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
idna==3.2
imageio==2.9.0
imagesize==1.3.0
imgaug==0.4.0
importlib-metadata==4.6.4
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib~=3.4.2
mistune==0.8.4
mmcv-full==1.3.16
mmdet==2.18.0
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.3
numpy~=1.21.0
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow~=8.4.0
printy~=2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil~=5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools~=2.0.2
pycocotools-windows==2.0.0.2
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn~=1.0.1
scipy==1.5.4
Send2Trash==1.8.0
Shapely==1.7.1
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
terminado==0.11.1
terminaltables==3.1.0
testpath==0.5.0
tifffile==2020.9.3
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yapf==0.31.0
zipp==3.5.0
